PC_Hub_Migration/requirements.txt
aiohttp==3.9.5
aiosqlite==0.20.0
pyserial==3.5
pyserial-asyncio==0.6
PyYAML==6.0.1
rich==13.7.1
python-json-logger==2.0.7
